+++ src/com/cpa/view/TestsPanel.java
@@ -38,15 +38,15 @@ public class TestsPanel extends JPanel {
   private void drawPoint(Vertex p, Graphics g, Color color) {
     Graphics2D g2 = (Graphics2D) g;
     g2.setColor(color);
-    g2.drawRect(p.x - 1, p.y - 1, 3, 3);
+    g2.drawRect((int) p.x - 1, (int) p.y - 1, 3, 3);
   }
 
   private void drawLine(Line line, Graphics g, Color color) {
     Graphics2D g2 = (Graphics2D) g;
     g2.setColor(color);
-    g2.drawLine(line.point.x, line.point.y, line.point.x + 1000, (int) (line.point.y
+    g2.drawLine((int)line.point.x,(int) line.point.y, (int)line.point.x + 1000, (int) (line.point.y
         + 1000 * line.director.y));
-    g2.drawLine(line.point.x, line.point.y, line.point.x - 1000, (int) (line.point.y
+    g2.drawLine((int)line.point.x, (int)line.point.y,(int) line.point.x - 1000, (int) (line.point.y
         - 1000 * line.director.y));
     drawPoint(line.point, g2, Color.RED);
     drawPoint(new Vertex(line.point.x + (int) line.director.x + 100, line.point.y + (int) line.director.y + 100 * (int) line.director.y), g2, Color.GREEN);
